\name{cfa}
\alias{cfa}
\title{Confirmatory Factor Analysis}
\usage{
cfa()
}
\examples{
\dontrun{
library(Statsomat)
cfa()
}
}
\value{
{Shiny app opens in viewer or browser.}
}
\description{
A Shiny app for automated Confirmatory Factor Analysis (CFA) based on the R package \href{https://lavaan.ugent.be/}{lavaan}. Single-group, first-order CFA for datasets up to 5000 observations, 25 (approximately) continuous variables and 5000 KB. An interpretation in natural language and the R Code to reproduce the results is included in the report. Run the app locally by calling the function or launch it directly in the web from \url{https://statsomat.shinyapps.io/Confirmatory-Factor-Analysis}. Follow the Instructions described in the GUI to use the app and generate a report. Check also the GitHub repository \url{https://github.com/Statsomat/CFA}.
}

