\name{corrana}
\alias{corrana}
\title{Correlation Analysis}
\usage{
corrana()
}
\examples{
\dontrun{
library(Statsomat)
corrana()
}
}
\value{
{Shiny app opens in viewer or browser.}
}
\description{
A Shiny app for automated Correlation Analysis for (approximately) continuous variables. An interpretation in plain English and the R Code to reproduce the results is included in the report. Run the app locally by calling the function or launch it directly in the web from \url{https://statsomat.shinyapps.io/Correlations}. Follow the Instructions described in the GUI to use the app and generate a report. Check also the GitHub repository \url{https://github.com/Statsomat/CORRANA}.
}
