\name{edar}
\alias{edar}
\title{Exploratory Data Analysis with R}
\usage{
edar()
}
\examples{
\dontrun{
library(Statsomat)
edar()
}
}
\value{
{Shiny app opens in viewer or browser.}
}
\description{
A Shiny app for automated Exploratory Data Analysis with R. Run the app locally by calling the function or launch it directly in the web from \url{https://statsomat.shinyapps.io/Descriptive_statistics/}. Follow the Instructions described in the GUI to use the app and generate a report. Check also the GitHub repository \url{https://github.com/Statsomat/edar}.
}
