\name{pca}
\alias{pca}
\title{Principal Components Analysis}
\usage{
pca()
}
\examples{
\dontrun{
library(Statsomat)
pca()
}
}
\value{
{Shiny app opens in viewer or browser.}
}
\description{
A Shiny app for automated Principal Components Analysis (PCA) based on the R package \href{http://factominer.free.fr/index.html}{factominer}. An interpretation in plain English and the R Code to reproduce the results is included in the report. Follow the Instructions on the webpage of the app \url{https://statsomat.shinyapps.io/Principal-components-analysis/} to generate the report. Check also the GitHub repository \url{https://github.com/Statsomat/PCA}.
}

